ArmPlatformPkg: CodeQL Fixes.
Makes changes to comply with alerts raised by CodeQL.

The issues here fall into the following category:

1. comparison-with-wider-type

+++ ArmPlatformPkg/Drivers/PL061GpioDxe/PL061Gpio.c
@@ -100,7 +100,7 @@ PL061SetPins (
   IN UINTN  Value
   )
 {
-  MmioWrite8 (PL061EffectiveAddress (Address, Mask), Value);
+  MmioWrite8 (PL061EffectiveAddress (Address, Mask), (UINT8)Value); // MU_CHANGE - CodeQL Change
 }
 
 /**
@@ -236,14 +236,14 @@ Set (
 
     case GPIO_MODE_OUTPUT_0:
       // Set the corresponding direction bit to HIGH for output
-      MmioOr8 (RegisterBase + PL061_GPIO_DIR_REG, GPIO_PIN_MASK (Offset));
+      MmioOr8 (RegisterBase + PL061_GPIO_DIR_REG, (UINT8)GPIO_PIN_MASK (Offset)); // MU_CHANGE - CodeQL Change
       // Set the corresponding data bit to LOW for 0
       PL061SetPins (RegisterBase, GPIO_PIN_MASK (Offset), 0);
       break;
 
     case GPIO_MODE_OUTPUT_1:
       // Set the corresponding direction bit to HIGH for output
-      MmioOr8 (RegisterBase + PL061_GPIO_DIR_REG, GPIO_PIN_MASK (Offset));
+      MmioOr8 (RegisterBase + PL061_GPIO_DIR_REG, (UINT8)GPIO_PIN_MASK (Offset)); // MU_CHANGE - CodeQL Change
       // Set the corresponding data bit to HIGH for 1
       PL061SetPins (RegisterBase, GPIO_PIN_MASK (Offset), 0xff);
       break;
